Add LogAbort which does an abort()
Also, if logging to a file, dump a stack backtrace into the file.

Also, within log_functions.c, revert to the original PTHREAD_RWLOCK macros
except in display_log_component_level, where we call the pthread_rwlock
functions directly and if we fail to lock, we change the log level to
NIV_ABORT, and if we fail to unlock, we simply abort.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>