Introduce state types for NLM state_t with hash table and related functions
To implement multiple file descriptors, we will need a place to hang the
file descriptor, so NLM will need the equivalent of a state_t. Basically
we will overload the NFS v4 state_t which will allow us to simplify some
SAL logic in a future patch. We will have NLM states for both locks
and shares.

We also add a state type for 9P. The 9P state serve as both a share AND
a lock state. Conveniently, the arrangement of state_lock and state_share
objects facilitate this. We define a new state_9p_fid structure to clarify
the fields we will use.

Change-Id: I1f7d51d80d8a5755318527593651c681d642f66c
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>