NFSv4.1/pNFS/Flexible Files: add Flexible Files Layout XDR definition
This commit brings the required changes to XDR related headers in order
to support the new Flexible Files Layout.
Due to (known) bugs in checkpatch.pl, this patch won't pass checkpatch
verification. Mostly, all of nfsv41.h (my changes but not only) will trigger
bunches of complains from checkpatch. Due to this, this patch is pushed
in --no-verify mode.

Change-Id: Idd2cdbd15fc61b6220a9093a6ba27f1250f9356b
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>