Nullfs : rewriting simple fsal functions using new handle
This commit contains the functions that can be easily corrected to use the
new nullfs_fsal_obj_handle type. They simply call the function of the underlying
subfsal with the appropriate op_ctx->fsal_export value.

In almost all of these functions, I check if attribute masks are cleared and
copy the attributes if not. I did that because I'm not sure if fsals and
inode exchange attributes only in getattr/setattr/xattr functions.

Change-Id: I06b2991e4a7ca1608ea545f8d0d8fac37613386d
Signed-off-by: Vincent Saulue-Laborde <Vincent.Saulue@bull.net>