Fix mount_work list corruption with dynamic add/remove export.
add_export_commit calls export_commit which adds the export to
mount_work list. The export is not removed from the list ever, but the
export itself is freed when it is removed. There is no need to add the
export to mount_work list from add_export_commit as it directly calls
mount_gsh_export to create the pseudo export.

Created export_commit_common function that has a flag to indicate if it
is called from export_commit or add_export_commit. An export is added to
work_list only when it is called from export_commit!

Change-Id: I6ebda8cc46877749151069c2811df6fb06d6457f
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>