Fix upcall directly changing attributes
This notification is completely asynchronous.  If we happen to change
some of the attributes later, we end up over writing those with these
possibly stale values as we don't know when we get to update with these
up call values. We should probably use time stamp or let the up call
always provide UP_TIMES flag in which case we can compare the current
ctime vs up call provided ctime before updating the attributes.

For now, just invalidate the attributes and let ganesha fetch attributes
as needed.

Change-Id: I22304f96409f1d854758d6143ce70d496ed0052c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>