Update patch set 8
Patch Set 8:

(2 comments)

I left a couple of comments regarding FSAL_PROXY.

I am just curious why we use 0x1010 and 0x2020 to flag endian? How about a single bit in a unint_8 flag? If set, it is little-endian, otherwise, it is big. I am a little cautious because we are talking about endian here, but the flag itself is wider than a byte.

Patch-set: 8
Label: Code-Review=0