Update patch set 11
Patch Set 11:

(1 comment)

Hi Frank and Marc,

I think this patch is good to PROXY as long as it drops the changes in xdr_nfs_fh4().

Because PROXY acts as a client to the back-end NFS server, file handle should be opaque to it. Therefore, it does not matter to PROXY if there is a flag inside or not.

Hope it helps.

Best,
Ming

Patch-set: 11
Reviewer: Ming Chen <1004505@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0