Update patch set 16
Patch Set 16: Code-Review-1

The fixups to NLM and FSAL_PSEUDO look good.

We still need some way to send the original file handle on NLM_GRANT.

Note that NLM_GRANT is the call back, so look at NLM_GRANT, not it's response (the response is just the client accepting the granted lock).

Patch-set: 16
Label: Code-Review=-1