Inode & servers layers : Checking attrlist when allocating fsal_obj_handle
This is a temporary solution to ensure that unpatched fsals can still work.
Whenever the cache_inode or the 9P/nfs servers allocate an handle, it checks
that the attrlist field is set. If it is not the case, they set the pointer
to the old attribute field of fsal_obj_handle.

Change-Id: I8a0fea20cad79ed4286d3d81990e750562a5fd7f
Signed-off-by: Vincent Saulue-Laborde <Vincent.Saulue@bull.net>