FYI - First pass of FSAL API to support multiple file descriptors
The multiple file descriptor implementation will accomplish several things.

First, the open_fd operation now allows atomic open, create, and set attributes
to the extent the underlying filesystem can. It also supports open by name or
by handle. Share reservations are also included here.

Second, the FSAL is provided a space to maintain separate file descriptors for
locks.

Note that Ganesha may also need to maintain a generic I/O file descriptor for
a file for anonymous I/O (a V4 client using a special stateid or a V3 client
doing any I/O when we can not associate that I/O with a share reservation or
lock - details here to be determined further in implementation).

The FSAL need not use a separate lock_fd if it is able to ride on top of the
open_fd and otherwise express a lock owner. In that case, the lock_fd will
basically never be opened.

Several calls allow passing the anonymous fd, the open_fd, and the lock_fd
to support the FSAL's use and depending on which are available to the
protocol layer when making a call.

It is expected that this interface will be refined and this patch updated.

Change-Id: Id6a734ee925cd343fb4b47d2ced3cdf19f270aff
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>