New FSAL API to support multiple file descriptors
The multiple file descriptor implementation will accomplish several
things.

First, the open2 operation now allows atomic open, create, and set
attributes to the extent the underlying filesystem can. It also
supports open by name or by handle. Share reservations are also
included here.

Second, the FSAL provides the mechanism to allocate state_t objects
allowing an FSAL to allocate additional space for it's needs. The
FSAL is sufficiently aware of state_t to be able to access
openstate for lock state_t's.

If the FSAL needs something similar for anonymous or general access
to an inode, it may allocate such space within it's private
object handle structure.

All calls still take an object handle and most all also take a state_t.

There is a new commit2 to allow different behavior when supporting
multiple file descriptors and different assumptions on who controls
file descriptors or whatever.

It is expected that this interface will be refined and this patch
updated.

Change-Id: Id6a734ee925cd343fb4b47d2ced3cdf19f270aff
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>