Use FSAL alloc_state to allocate state_t objects
Also allow for state to already be allocated for state_add_impl. This is
to support NFS v4 OPEN where we need the state to be able to open the
file but are not ready to add the state to the hash table yet.

Also call close2 before free_state so we don't leak file descriptor
resources.

Change-Id: I312b26f1abce2cd08a570734b33b9585d153848c
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>