Fix compile problem caused by wrong cephfs lib and header path setting
The following message was output during compiling the ceph module on my x86_64
machine:
  /usr/bin/ld: skipping incompatible /usr/lib/libdl.so when searching for -ldl
  /usr/bin/ld: skipping incompatible /usr/lib/libdl.a when searching for -ldl
  /usr/bin/ld: skipping incompatible /usr/lib/libpthread.so when searching for -lpthread
  /usr/bin/ld: skipping incompatible /usr/lib/libpthread.a when searching for -lpthread
  /usr/bin/ld: skipping incompatible /usr/lib/librt.so when searching for -lrt
  /usr/bin/ld: skipping incompatible /usr/lib/librt.a when searching for -lrt
  /usr/bin/ld: skipping incompatible /usr/lib/libc.so when searching for -lc
  /usr/bin/ld: skipping incompatible /usr/lib/libc.a when searching for -lc

It is because we set CEPH_PREFIX to "/usr" if we didn't input a specified
ceph install path, and when we executed the cmake file of the ceph module,
we though CEPH_PREFIX was set, and add new lib path (${CEPH_PREFIX}/lib) for
ld, then the above problem happened. In fact, The ceph lib and header file
path has been found and set during the global cmake process, we needn't set
it again, so clean up them.

Change-Id: I30d3887a7bd008d272fbc628dbc9847165a0b8d5
Signed-off-by: Miao Xie <miaoxie@huawei.com>