Error handling: handle FSAL's ENODATA (xattr)
This change is mainly for 9P since NFS doesn't handle xattrs.
The previous commit makes that we do not actually need all the
conversions, but we still get a lot of lines like this:
posix2fsal_error :FSAL :CRIT :Mapping 61(default) to ERR_FSAL_SERVERFAULT

NFS error conversions still return serverfault on posix error ENODATA
as it used to.

Change-Id: I440916e76e25b3c9b8af726dc92e4483315e7f35
Signed-off-by: Dominique Martinet <dominique.martinet@cea.fr>