lock: Fixed an issue with lock_length while splitting locks
In 'subtract_lock_from_entry()', while doing the right split of
the locks, if the 'found_entry_end value' is UINT64_MAX, it
indicates that the 'sle_lock.lock_length' is zero and the lock is
held till the end of the file. In such cases the split-lock length
should be zero as well to indicate the file end.

Change-Id: I69c1061e7b81634db12e0efcd210637aee7d79ef
Signed-off-by: Soumya Koduri <skoduri@redhat.com>