If dirent is stale in cache_inode_find_by_name don't leak ESTALE
If the dirent has a stale object handle, we want to invalidate the dirent
(well, we invalidate the whole directory cache), and continue to lookup
or open since there might be a NEW object with that name.

If CACHE_INODE_NOT_FOUND is returned (which would be the case for an
open) we do the same thing.

Change-Id: I748701fb267f06ac8241f8c6fd60984982fc78fc
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>