build: make it possible to disable FSAL_NULL
When FSAL_NULL gets build, there will be an additional
nfs-ganesha-nullfs RPM generated. To prevent any confusion if this
package is needed or not, add an option to disable the building of
FSAL_NULL and don't build the RPM package in that case.

The usage is similar to other FSALs, FSAL_NULL is enabled by default,
disabling cam be done like this:

  $ cmake -DUSE_FSAL_NULL=NO ... /path/to/nfs-ganesha/src/

Change-Id: Ifc932fd35b195bc1ea33dd2b899b4d1bab0dbab3
Signed-off-by: Niels de Vos <ndevos@redhat.com>