cache_inode_lru: shift object addr in lru_lane_of
The goal is to reduce skew in the mapping of objects to
lanes, reducing contention for lane locks.  A change similar to
this was found to reduce lock contention in a cognate code in
Ceph's Accelio transport.

Marc Eshel reported no measureable difference with and without
this change, on a workload that was seeing high lock contention
with 1.5M LRU entry limit.

Change-Id: I492f165fbef5938187899fead6427a68dc7c3fa7
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>