WIP - New cache inode wrappers for FSAL read2, write2 and commit2
Now that handling of the file descriptor (or whatever) if any is
handled by the FSAL, these functions become much simplere. Particularly
the use of the content_lock drops out (the FSAL will do it's own
protection of the file descriptor or whatever.

We also get rid of paranoia code correctness checks.

Change-Id: I42de289ca7ada24e33aec06da4f7bbf98719b94a
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>