Fix incorrect atime/mtime with EXCLUSIVE4/4_1 create
The client was creating files with EXCLUSIVE4 method, and ganesha server
uses atime and mtime metadata fields to store the provided verifier.
According to the RFC paragraph below, the server needs to notify the
client about what metadata fields it used to store the verifier so that
the client can later send a setattr to update those fields. We
don't do that, so the client never updates atime or mtime. In the end,
we end up files with bogus atime (and mtime).

Just mounting a ganesha export in NFSv4.0, and then doing "cp
<local-file> <new-file-in-NFS-mount>" will reproduce the issue and can
be verified with 'stat' command.

And here is the RFC paragraph from RFC 7530:

If the server uses an attribute to store the exclusive create verifier,
it will signify which attribute was used by setting the appropriate bit
in the attribute mask that is returned in the results.

RFC 5661 has a similar note for EXCLUSIVE4_1 as well, so added setting
atime and mtime attribute flags in the response attrset mask.

Also fixed sattr.mask when sattr.mode is set.

Change-Id: Ic0a0d4af9d13268d574b79dedc269a7f67a17872
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>