Use CMake's internal concept of target dir
It turns out that ubuntu has some 64 bit systems with lib and lib64, and
some with lib and lib32.  This breaks with our current bitness checking.
CMake, howeevr, has a global property that knows this.  Attempt to use
this property.  This should have no effect on systems that use lib and
lib64.

Change-Id: I550bad93be4e3973444463d3f1ab7e9f0901d93c
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>