service files: Provide an option to pass EPOCH value
Provide an option to generate and pass EPOCH value while
starting ganesha.nfsd.

$EPOCH environment variable contains the epoch value parameter
where as $EPOCH_FILE represents the file where the $EPOCH
value is copied into.  '/var/lib/ganesha/epoch' is used as
default path. Changes to this file location have to be updated in
'/etc/sysconfig/ganesha' and SysV-init/systemd-unit scripts of
nfs-ganesha service.

The contents of this $EPOCH_FILE should be in the following format
EPOCH="-E <epoch_value>"

Change-Id: Ie090959fb35fee0fd532289fbe80863a183cabc9
Signed-off-by: Soumya Koduri <skoduri@redhat.com>