Add a file with some helgrind error suppressions
There are a bunch of libntirpc errors that should be examined.

There are some errors in Ganesha core code that should be examined
but it's handy for now to suppress the errors.

helgrind doesn't handle atomics well, so this file excludes all
the ones I've hit. Intrinsicly they are safe so no need to have
any context surrounding their suppression.

Change-Id: Ieee5a2e57bdfae98c78fa67e03d096508f65e02f
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>