service files: Provide an option to pass EPOCH value
Provide an option to generate and pass EPOCH value while
starting ganesha.nfsd.

'/var/lib/ganesha/epoch' is used to copy the EPOCH value.
The contents of this file should be in the following format
      EPOCH="-E <epoch_value>"

A new service "ganesha-epoch" is used to generate
and copy EPOCH value to '/var/lib/ganesha/epoch'.

Note: This new service and EPOCH file are optional but need to
be defined by FSAL if required to change default EPOCH value.

Change-Id: Ie090959fb35fee0fd532289fbe80863a183cabc9
Signed-off-by: Soumya Koduri <skoduri@redhat.com>