FSAL_CEPH: NUL-terminate symlink buffers
The ceph_ll_readlink() function returns the number of bytes copied
on success, so we can copy and NUL-terminate the buffer being returned
to Ganesha without a length check.

Introduces a strdup-like wrapper for strlcpy that takes a bound, to
avoid the additional NUL-scan we were previously doing.

Fixes a defect found and fixed by wangsongbo <wangsongbo@unitedstack.com>,
but avoids the call to memset() on the returned buffer.

Change-Id: Iac27c9ca6bcf99a762055355e60212d558d36938
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>