Avoid holding so_mutex while destroying the lock.
uncache_nfs4_owner() may result in destroying the so_mutex and freeing
the state owner when reap_expired_open_owners() calls it. In this case,
so_mutex lock must not be held as it may get destroyed. If callers of
uncache_nfs4_owner have additional reference on the state, they can call
uncache_nfs4_owner() while holding the so_mutex though as the lock is
guaranteed to be there after the call.

Change-Id: Ib06b56fc8e5332424f829de6b7241a0a46fa0a0c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>