[WIP] FSAL_GLUSTER: Add support for SEEK procedure
This makes it possible to detect sparse areas in files. The SEEK
procedure is part of NFSv4.2 and a description of the functionality is
described here:
  https://tools.ietf.org/html/draft-ietf-nfsv4-minorversion2-41#section-15.11

The basic infrastructure was already in place, FSAL_GPFS seems to
support SEEK too.

Gluster gained support for SEEK_DATA/SEEK_HOLE with version 3.8
(currently not yet released). Just like with NFS, the actual position in
the file-descriptor with lseek() is kept client-side because the READ
and WRITE procedures pass an offset.

The glfs_seek() function in glusterfs < 3.7.9 is broken, it does not
validate the 'whence' parameter correctly. A fix has been included in
version 3.7.9 to return EINVAL (http://review.gluster.org/13631). To use
SEEK with Gluster, gfapi should at least contain this change. Updating
the version in CMakefile.txt to reflect it.

Change-Id: I9d9474212c6853deb3390a8154e03973aa8519a4
Signed-off-by: Niels de Vos <ndevos@redhat.com>