VFS FSAL fsid_type configuration parameter is not parsed correctly.
Two issues -

1. One major issue is that the default value is set to -1 (since the
parameter type is u32 the actual default value set is 0xffffffff). Now
the do_block_load() for CONFIG_ENUM does the following to set the
export fsid_type -
	*(uint32_t *)param_addr |= num32;

For example, if "Fsid_type = Two32" is specified in configuration,
num32 will be 4, but the operation above does not set the correct value
as seen below in the log -

    839 :do_block_init :CONFIG :F_DBG :0x7ffff4695250 CONFIG_ENUM fsid_type mask=00000000 def=ffffffff value=ffffffff
    ...
   1200 :do_block_load :CONFIG :F_DBG :0x7ffff4695250 CONFIG_ENUM fsid_type mask=00000000 flags=00000004 value=ffffffff
   1225 :do_block_load :CONFIG :F_DBG :0x7ffff4695250 CONFIG_ENUM_SET fsid_type mask=00000000 flags=00000004 value=ffffffff

The fix I submitted changes the default value to zero (0), so that the
above will work, but I am not sure that that is correct way to fix it.

2. Another issue is that a "break" statement is missing in-between the
cases in the switch block. See the above logging.

Change-Id: I76a97210a471c1f4071d132ec13a23a032907192
Signed-off-by: Krishna Harathi <kharathi@exablox.com>