Add FSAL_MDCACHE
Add metadata caching as a stackable FSAL (named FSAL_MDCACHE).  This is
largely the cache_inode code translated into the FSAL API.  However, it
simplifies some things, and normalizes API style and access.

Among other changes, this converts FSAL_UP to allow stacking as well,
resulting it the signals that matter being handled by the layers that
care, and the rest can propagate further up the stack.  This will allow
for a DATA caching layer in the future.

Change-Id: I4ac40cf7cb2776bda7c2cf8333a53c4498278f91
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>