Streamline fsal_close
The same method for fsal close is used whether support_ex is enabled or
not. This patch makes a simple fsal_close function that handles all the
conditions. Since sometimes, support_ex has already been checked and is
negative, a flag parameter is used to indicate if checking should be
skipped (in which case it is assumed negative).

This also assures that legacy fsal close method is only called when
fsal status call indicates the file is open.

A call to fsal_close for a non-regular file simply results in a
successful return.

Change-Id: I907f5fc8b4aa2462395186d8c10592e7b16567f5
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>