Clean up remove/rename paths
The remove/rename paths did several lookup and getattr pairs that are
really unnecessary.  Since most files have only one link, assume that an
overwrite rename and a remove actually remove the file, and drop them
from the cache, as long as they have no open state.  This requires
tracking when the last state goes away, so cleanup can be performed
then.

This re-exposed an issue with recursive locks when cleaning state.

State cleanup can result in a recursive lock.  This was handled before
by a flag passed to cache_inode_lru_unref().  Until state refcounting
can be implemented, add an no_cleanup flag to state_hdl that is
protected by the state_lock, and use it to indicate when cleanup should
not occur.

Change-Id: Id96181964250d0663add2f67e479d768aa4bb96c
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>