[RFC] EEXIST error due to mdcache_avl_qp_insert collision
This problem is observed with V2.3-stable, but this may not be the righ
fix for V2.4. Let us discuss if this problem really exists in V2.4, if
so, what would be the right fix.

We have seen spurious EEXIST error on directory/file creations due to
mdcache_avl_qp_insert() collision detection. This seems to have been
caused by FSAL asynchronous upcalls that invalidate directory entries and
a subsequent nfs4_op_readdir() call racing with mdcache_inode_create()
call.

The object is created but before the create thread adds it to the
directory cache, nfs4_op_readdir thread adds it causing collision
detection in create thread.

mdcache_dirent_add() is now made to return success on failure
to insert a directory entry due to name collision.

Change-Id: I1d0a9e575f31b888a2f5a025e72e8103c193150a
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>