Change behavior - put_ref and get_ref are required.
When FSALs are detached from MDCACHE, they MUST implement get_ref and
put_ref. Various bits of export and state code depend on holding a ref
to an obj_handle in order to maintain state. This state used to be
attached to a cache_inode entry, but with MDCACHE now optional, there
MUST be a way to retain these obj_handles.

Change-Id: Ib80f07e2d79ac643f05baa4f57efa103135345e8
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>