Convert attributes to return by copy
Known issues that should have been addressed:

- Complete audit of ACL handling
  - Callers passing in attributes to be set should call
    fsal_release_attrs to clean up any ACL passed in.
  - Setattr must make sure a replaced ACL in MDCACHE is properly handled
  - The same with getattr

- While most FSALs use struct stat to create a new fsal_obj_handle,
  FSAL_GPFS uses a struct attrlist. This must be audited for proper cleanup
  and if there are any other places that should try and access the cached
  attributes.

- FSAL_GLUSTER should be checked

- FSAL_GPFS may have some attribute cleanup yet to be done

- The only FSALs I have tested are MDCACHE, VFS, CEPH, and PSEUDO

- More testing

----- final text for commit
The attributes are no longer cached in the fsal_obj_handle or FSAL's
private handles, except by FSAL_MDCACHE. This allows for better management
of the attributes, with a copy out to the caller when necessary, which also
allows for a consistent set of attributes to be used by the caller.

The copy out is not too expensive, the only big attribute is an ACL which
is copied by reference (which thus requires the callers to release the
potential reference when done with the attributes). There is allowance for
specifying which attributes are desired which may allow for simpler getattr
in the future.

fsal_prepare_attrs should be called before making a call to getattrs. This
allows any future changes to struct attrlist that might require
initialization to be hidden from the caller.

fsal_release_attrs MUST be called when done with a set of attributes. It
will assure that anything not statically part of struct attrlist is
properly released (currently this releases the reference on an ACL). If in
the future additional referenced items are added to attrlist, they will
quietly be taken care of.

Future Plans for MDCACHE:

Maintain a separate validity flag for the ACL, this would allow only
fetching an ACL when necessary. ACLs might even be able to have a longer
valid period (since they don't change too often).

Change-Id: Ibf26a349aee29fba362a3cde636778f181c6a8cf
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>