Pass attributes out on fsal_obj_handle create and readdir
When an FSAL object is created, most FSALs must get the file attributes
in order to properly create the fsal_obj_handle, in this case, since many
callers (including MDCACHE) will need attributes, it makes sense to pass
them up instead of having an additional getattrs call.

The same can be said of readdir, so go ahead an pass the attributes up
on the call back.

MDCACHE has also been somewhat re-organized to perform cached object
creates in the same way as much as makes sense.

Change-Id: Ia89bb16356fa5117169e80b66b7d27e0a6a0e23e
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>