MDCACHE: Add call to FSAL merge method back in
This late addition to cache_inode didn't make the transition to MDCACHE.

If open2 is by name and thus instantiates a new object, which thus results
in a call to mdcache_new_entry and subsequently results in a race with
another call, we must merge the object handles (to merge the share
reservations). This might fail, or there might be some other failure during
instantiation of the new mdcache_entry, in which case we must close the
recently opened file state.

Change-Id: I95307233e57e3d8e07515a35d2cddc11177f8d24
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>