MDCACHE: Resolve name collisions in dirent tables
If there is a name collision on adding a dirent to a directory, and the
existing dirent has the same handle hash key and the same flags, then
resolve as a succesfull addition.

This also changes the interface to mdcache_avl_qp_insert to take the dirent
pointer by reference so that it can be updated in the case of a collision,
or released and set to NULL in the case of any error.

The callers never actually use the dirent pointer after calls to
mdcache_avl_qp_insert, however, this interface change guarantees we will
catch any unexpected uses.

Change-Id: Ia2e336204ee3882c14c2112e1282392549f07f83
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>