Allow privilegedport to be true in export_defaults block
Things defined in export_defaults block are applied to all exports
including the pseudo export. If someone sets privilegedport to true in
the export_defaults block, then it is not possible to allow any
individual export to be accessible with non-privilegedport! In other
words, the following config won't work today. Clients trying to mount
"/export1" with non-privilegedport will fail as putrootfh will
fail with NFS4ERR_ACCESS error.

	export_defauls { privilegedport = true; }
	export { path = /export1; privilegedport = false; }

We should set the pseudo export with privilegedport to false, and let
the value in export_defaults block apply only to regular exports.

Change-Id: Ic499b3f7199ab212e66b851e0dd8b0b0406782ca
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>