Make sure we use anonymous uid/gid from export_perms or EXPORT_DEFAULTS
We want to make sure we use a configured anonymous uid or gid anywhere
possible.

If op_ctx->export_perms is available use that, otherwise use the value set
in EXPORT_DEFAULTS config block if any, and only if that was not set, use
the code defaults of ANON_UID/ANON_GID (-2).

To this extent, export_check_access() may be called with no export set
and get_anonymous_uid() and get_anonymous_gid() will fetch the most
appropriate value.

So, the places this applies:

The order of callss in nfs_rpc_execute have been rearanged so that
we can rely on export_check_access() to set the correct initial
export_perms.

When decoding an ACL, owner, or owner_group attribute, we call
get_anonymous_uid() or get_anonymous_gid() as appropriate. This is
actually overkill since export_perms SHOULD be set anytime these
attributes are being decoded.

In principal2uid(), we use get_anonymous_uid() or get_anonymous_gid()
as appropriate. This should also be overkill.

In build_default_root() we don't bother setting the anonymous_uid and
anonymous_gid values since the options are not marked set and thus
when the export is used, it will pick up values from EXPORT_DEFAULTS
if they were set there, or code defaults otherwise.

In init_credentials() just use the values from op_ctx->export_perms since
they have already been set by a call to export_check_access(). Note that
clean_credentials() may set these to values based on the last used
export, however, the credentials are about to be discarded.

Change-Id: I9839f167c402edb968715bbc4ee6fb4fdd219d11
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>