gpfs: Fix incorrect return value from gpfs_ganesha when open fails
Currently it returns ENOSYS, and callers don't expect such a value.
What callers expect depends on the ioctl command they pass. It could return
-1 with errno set to ENOSYS, but there is no point in continuing in such
an environment. The current code tries to log an error to stderr, but
that goes to /dev/null in daemon mode. Replaced the printf with
LogFatal. Observed this open syscall failure in some SELinux enabled
environments.

Change-Id: I60857fea2198b6830342a58a09d95bce3ef0525c
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>