cmake: find libhandle.so on Debian Stretch and Ubuntu Xenial
xfslibs and xfslibs-dev on Stretch and Xenial install libhandle.so.1*
and libhandle.so in /lib instead of the more usual places, e.g.
/usr/lib/libhandle.so or /usr/lib64/libhandle.so. (Perhaps this is a bug
in Stretch's and Xenial's xfslibs packages?)

To make matters worse, libhandle.a on these distributions is installed
in /usr/lib/, with the result that cmake's check_library_exists() will
find them, but later on build will try to link with libhandle.a and fail
due to the mismatch between PIC code in the FSAL and the non-PIC code in
libhandle.a.

Thus we check first (order is important here) for the odd location, and
if found, set things up to link with it. Otherwise check the normal
location(s) and let the link find it the way it usually does.

Change-Id: I2deb24176d94bc68c9c4cca27b94cedc8bce68da
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>